[TRTLLM-7027][feat] Fuse d2t to logitsBitmaskKernel and fix a race condition in one-model spec (#7481)

+++ tensorrt_llm/_torch/models/modeling_speculative.py
@@ -177,7 +177,7 @@ class Eagle3DraftModel(DecoderModel):
 
         if config.draft_vocab_size is not None and config.vocab_size != config.draft_vocab_size:
             self.d2t = nn.Parameter(torch.empty((config.draft_vocab_size, ),
-                                                dtype=torch.int64),
+                                                dtype=torch.int32),
                                     requires_grad=False)
 
         if self.hidden_size_in != config.hidden_size:
